No longer passing google as a prop
Passing google around as a prop was creating a lot of extra code. Since it the google variable is global it's not really necessary. We were using the prop to check if google was loaded in some situations but it's a lot easier just to stop any of the code from being loaded until google is loaded. There doesn't seem to be any penalty in terms of performance by doing it this way. We are also no longer passing the geocoder in the payload each time we want to geocode. We just set it in the geocode module after google is loaded instead.