Refactored GoogleMap
Changed viewport prop to bounds because we can pass in either a bounds object or a viewport object, call fitBounds object on it, and the result is the same, so they are at least functionally equivalent. Also making it so that setCenter and fitBounds separately react to changes in the props. Using fitBounds seems to actually be all we need, but keeping location too in case there's something I'm missing.