Fixed bug where clicking inside the dropdown menu was closing it
Putting the click event on the button itself, instead of the main container, stops it from triggering toggleMenu method when you click inside the menu, since clicking anywhere in the menu triggers a click event that bubbles up to the parent. Adding the click event to the button inside ContainedButton allows us to listen to the click event from the button on the component with @click.