Putting fix for AutoComplete dropdown back in place
Making the search field icon an actual submit button seems to have brought back this bug where the form is submitted when you hit the enter key, which causes the search to be performed twice. I think it was happening previously because it was triggering a click even on the search button I had. When I replaced that button with an icon the issue went away. Now that it's a button again the bug is back. A better long-term solution will be to create our own dropdown instead of using Google's.