Fixed bug with location
We're calling this.geocoderResult.location.lat for the center_lat in the getGeoLayer action, so location needs to be an object. We could do this.geocoderResult.location.lat() instead of that but we might as well just store it as a LatLngBoundsLiteral since we don't need any of the class methods for it.