Changed search to filter listings by bounds
When the user drags or zooms the map, we want to filter listings that are outside the map bounds. Previously we were doing this by making a new call to the listing service. Now we just filter the listings we already have by checking their coordinates against the map bounds. It's working well if the amount of listings we have is beneath the cluster threshold because that means we have all of them. If we are clustering a large amount of listings we will still need to get some listings when listings in the search results go outside of the bounds because we don't have all the listings. We only have 20.