Created modifyParam class for search params
Created class to modify search params that need changes before being sent to the listing service. It maps search param names to class methods with the same name that return an object with necessary modifications. Using it to start including user_lat & user_lon when sort_by values are for distance.