Finished MapTypeControl
Refactored map controls as components that are positioned on top of the map. The Google Maps API for creating custom controls really wasn't fitting well at all. They want a DOM element to be created and passed to the googleMap.controls array. Creating a component programmatically is possible with Vue but passing it params and having them be reactive doesn't work well at all, and I definitely didn't want to stick with the paradigm of manually creating DOM elements from the examples in the Google Maps docs.