Fixed bug with map moving after boundary removed
Resetting the geoLayerCoordinates used for the boundary polygon back to an empty array was causing the map to snap back to it's original location after being moved, instead of staying in place and just removing the polygon. Setting it to not be visible prevents this from happening and the user experience is otherwise the same. Keeping the data around in case we wanted to be able to turn the boundary back on seems like a good idea anyway.