Created component to programmatically add components to google maps controls
It works to add components to the map using the Google Maps API map controls array as specified in their docs. However, I'm not completely clear on why this works, or if it would have any drawback. Looking at Zillow and Trulia, they don't seem to try to do it this way, so it seems like we shouldn't bother. Just committing this to keep an example around in case I change my mind.