Remove bounds for new searches & search immediately
Changes to the lates version of the Google Maps API have made it so that we can no longer rely on the "idle" event being triggered after calling map.fitBounds(). We were relying on this to make sure the map had finished updating before we do the listing search. Now, if the location hasn't changed, but the user initiates a new search in the same location, the listings are reset but the subsequent search is never triggered by the "idle" event that used to always fire. It appears that this event no longer fires if fitBounds() did not adjust anything, which it wouldn't since the location did not change at all.

The solution is to not wait for the idle event to fire when initiating a new search. Instead we do the new search immediately. This requires that we do not include the map bounds in the new search request. The bounds may be for the old location and therefore would limit the listings that are returned in a way that would be incorrect.